--- src/main/java/com/luti/auth/entity/Route.java
@@ -1,32 +1,0 @@
-package com.luti.auth.entity;
-
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "route")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Route {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="route_id", updatable = false, nullable = false)
-    private Long routeId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", referencedColumnName = "user_id", insertable = false, updatable = false)
-    private User userId;
-
-    @Column(name = "route_title")
-    private String routeTitle;
-
-    @Column(name = "route_content")
-    private String routeContent;
-
-}
