+++ src/main/java/com/luti/mypage/entity/Route.java
@@ -1,0 +1,41 @@
+package com.luti.mypage.entity;
+
+import com.luti.auth.entity.User;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "route")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Route {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "route_id", updatable = false, nullable = false)
+	private Long routeId;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", referencedColumnName = "user_id", insertable = false, updatable = false)
+	private User userId;
+
+	@Column(name = "route_title")
+	private String routeTitle;
+
+	@Column(name = "route_content")
+	private String routeContent;
+
+}
